Restructure test layout.
